pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "us-weather-energy-pipeline"
version = "0.1.0"
description = "US Weather and Energy Analysis Pipeline"
authors = [{name = "Eno-obong Etim Bassey", email = "enobongetimbassey101@gmail.com"}]

dependencies = [
    "pandas",
    "requests",
    "pyyaml",
    "streamlit"
]

[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "pytest",
    "mypy",
    "ruff"
]

[project.scripts]
run-pipeline = "src.pipeline:main"
data_fetcher = "src.data_fetcher:main"


[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]

[tool.mypy]
python_version = "3.9"
ignore_missing_imports = true

[tool.ruff]
line-length = 88

[tool.setuptools]
packages = ["src"]